wrong cast for 64 bit targets
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@27473 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: fc5693b8a944517de11e6c2bb8467d1dbcba45de

+++ src/router/upnp/src/gena.c
@@ -217,7 +217,7 @@ notify_prop_change(UPNP_CONTEXT *context, UPNP_SUBSCRIBER *subscriber)
 
 	/* set non-blocking IO to have connect() return without pending */
 	on = 1;
-	if ((ioctl(s, FIONBIO, (int)&on)) == -1) {
+	if ((ioctl(s, FIONBIO, &on)) == -1) {
 		upnp_syslog(LOG_ERR, "cannot set socket as non-blocking IO");
 		goto error_out;
 	}
